Example of breaking InterMine API models into multiple files.
